+++ docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
-## 0.14.0 - 2022-04-03
+## Unreleased
+- Breaking change: add [react-native-gesture-handler](https://docs.swmansion.com/react-native-gesture-handler/docs/installation/) v2 and [react-native-reanimated](https://docs.swmansion.com/react-native-reanimated/docs/fundamentals/installation) v2 as peer dependencies.
+## 0.16.0 - 2022-06-04
+- Improve performance [#210](https://github.com/hoangnm/react-native-week-view/pull/210)
+## 0.15.0 - 2022-05-27
+- Add scrollToTimeScrolled callback [#208](https://github.com/hoangnm/react-native-week-view/pull/208)
+## 0.14.0 - 2022-05-24
 - Add scrollToTime method [#205](https://github.com/hoangnm/react-native-week-view/pull/205)
 ## 0.13.0 - 2022-04-03
 - Support orientation changes [#201](https://github.com/hoangnm/react-native-week-view/pull/201)
@@ -20,7 +26,7 @@
 ## 0.7.0 - 2021-11-11
 - Add isRefreshing and RefreshComponent props [#112](https://github.com/hoangnm/react-native-week-view/pull/112)
 - Fix events break when changing hoursInDisplay [#133](https://github.com/hoangnm/react-native-week-view/pull/133)
-- Allow to drag-drop event [#143](https://github.com/hoangnm/react-native-week-view/pull/143) 
+- Allow to drag-drop event [#143](https://github.com/hoangnm/react-native-week-view/pull/143)
 - Support react-native above v0.59.0 only
 ## 0.6.1 - 2021-08-22